abstract Signup classes and implement stripe Signup stub
