+++ src/Tickets/Commerce/Gateways/Stripe/Signup.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace TEC\Tickets\Commerce\Gateways\Stripe;
+
+use TEC\Tickets\Commerce\Gateways\Abstract_Signup;
+
+class Signup extends Abstract_Signup {
+
+	/**
+	 * @inheritDoc
+	 */
+	public static $signup_data_meta_key = 'tec_tc_stripe_signup_data';
+
+
+}
